Clean and tidy up the `pkg/` dir (#4883)
* Fix "redundant type from array" warning


* Clean pkg/auth dir


* Clean more pkg/*


* Move pkg/response to kubeops


* Move pkg/http-handler to kubeops


* Move pkg/auth to kubeops
